requirements.txt
#https://download.pytorch.org/whl/cpu/torch-2.1.2%2Bcpu-cp39-cp39-linux_x86_64.whl

requests
feedparser
opengraph_py3
python-dotenv
requests
feedparser
tweepy
praw
attrdict
openai
#tts
Pillow
pyyaml


#./lazycommon
#./feedreaper
#./lazysummary
#./lazysocials
